% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{add_annotations}
\alias{add_annotations}
\title{Add a user-specified list of cluster annotations to a Seurat object}
\usage{
add_annotations(
  seurat_obj,
  annotations_df,
  cell_types_col = "CellType",
  relabel = TRUE,
  relocate = TRUE,
  alphabetize = TRUE,
  clusters_col = "seurat_clusters",
  annotations_col = "annotated_clusters"
)
}
\arguments{
\item{seurat_obj}{The Seurat object to annotate.}

\item{annotations_df}{Data frame containing cluster-to-cell-type mappings, typically
with "Cluster" and "CellType" columns.}

\item{cell_types_col}{The name of the column containing the cell type annotations.}

\item{relabel}{Whether to update the Seurat object's active identities. Sometimes
you just want to add the metadata.}

\item{relocate}{Whether to relocate the annotation column in metadata.}

\item{alphabetize}{Whether to alphabetize the cell types.}

\item{clusters_col}{The name of the column containing cluster IDs.}

\item{annotations_col}{The name of the new metadata column for the annotations.}
}
\value{
A Seurat object with added annotation information.
}
\description{
This function adds a user-specified list of cluster annotations to the Seurat object.
It can be used for both manual and automated annotation, updating cell identities
and adding annotation columns to the metadata.
}
