% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airr.R
\name{add_family_info}
\alias{add_family_info}
\title{Add in family and gene information from \code{alakazam.}}
\usage{
add_family_info(combined_airr)
}
\arguments{
\item{combined_airr}{An AIRR-formatted data.frame.}
}
\value{
A data.frame with six new columns containing gene family and gene information.
}
\description{
Extracts and adds V, D, and J gene family and gene information to an AIRR-formatted
dataframe using the alakazam package functions.
}
\details{
biomaRt also has a getGene function, so we have to be specific.
}
