% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_info_bar}
\alias{add_info_bar}
\title{Add an information bar on top of a given ggplot}
\usage{
add_info_bar(
  plot,
  method = "contains",
  info_type,
  info,
  sort = TRUE,
  text_size = 8,
  label = TRUE,
  angle = 0,
  top_side = TRUE
)
}
\arguments{
\item{plot}{A ggplot2 object to which information bars will be added.}

\item{method}{The method for getting the info bar information; one of \code{add}, \code{contains}, \code{join}.}

\item{info_type}{What to add e.g. Dataset, Cell_Type, etc.}

\item{info}{Data frame containing the information to join/add, must contain
the same column as plot$data to join by.}

\item{sort}{Sort the info alphanumerically.}

\item{text_size}{The size of the text in the bars.}

\item{label}{Whether or not to include what you are adding info for (good for datasets).}

\item{angle}{Rotation degree.}

\item{top_side}{Plot on top (by default) or on the right side.}
}
\value{
A ggplot plot.
}
\description{
This function adds informational strips (facets) to the top of an existing \code{ggplot2}.
}
\details{
This is very useful for splitting by dataset, showing cell types, etc.
\code{contains} is not really needed.
Note that providing a named list of features to a DotPlot will automatically group them without the need for this function.
}
\examples{
\dontrun{
p <- ggplot(data, aes(x, y)) + geom_point()
add_info_bar(p, info_type = "Dataset", info = dataset_info)
}
}
