% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athanor-package.R
\docType{package}
\name{athanor-package}
\alias{athanor}
\alias{athanor-package}
\title{athanor: Integration of Single-Cell Gene Expression and B Cell Receptor Data}
\description{
Integrates paired single-cell transcriptomics and B cell receptor sequencing data for improved B cell subtyping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eba28/athanor}
  \item Report bugs at \url{https://github.com/eba28/athanor/issues}
}

}
\author{
\strong{Maintainer}: Edel Aron \email{edel.aron@yale.edu} (\href{https://orcid.org/0000-0002-8683-4772}{ORCID})

}
\keyword{internal}
