% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{automated_annotation}
\alias{automated_annotation}
\title{Run automated cell type annotation}
\usage{
automated_annotation(
  seurat_obj,
  annotation_method,
  reference = "pbmcref",
  majority_voting = FALSE,
  over_clustering = NULL
)
}
\arguments{
\item{seurat_obj}{The Seurat object. Must be the path to a H5AD object if using CellTypist, or a Seurat object if using Azimuth.}

\item{annotation_method}{Which method to use: CellTypist or Azimuth}

\item{reference}{Reference or model to use for prediction.}

\item{majority_voting}{Whether to enable majority voting for CellTypist predictions, which refines predictions based on local cluster information but may increase runtime.}
}
\value{
A data.frame with the annotations for each cell
}
\description{
This function runs automated cell type annotation using \code{CellTypist} or \code{Azimuth}.
}
\details{
Supports CellTypist and Azimuth annotation methods.
Assumes that the \code{Cells()} of \code{seurat_obj} are properly formatted (i.e. unique).
For \code{CellTypist}, assumes the H5AD file and predictions have already been generated.
For \code{Azimuth}, requires the \code{AzimuthAPI} package (not a dependency of this package,
since it is not on CRAN), which runs predictions through Satija Lab's cloud API.
See \url{https://github.com/satijalab/azimuth} for installation instructions.
This would typically be used after \code{\link[=seurat_pipeline]{seurat_pipeline()}}.

Even though CellTypist runs upon a scanpy object, we call the first parameter \code{seurat_obj} for consistency with the rest of the package. It should be the path to a H5AD object if using CellTypist, or a Seurat object if using Azimuth.

"Majority voting refines the prediction result in a local cell cluster by
choosing the dominant cell type label but may increase the runtime especially
for a large dataset due to the over-clustering step. This approach usually
improves the cell annotation, as voting is conducted in small subclusters
derived from over-clustering (cells belonging to a given cell type will be
assigned the same label regardless of potential batch effects separating them)."
\itemize{
\item https://www.celltypist.org/tutorials/onlineguide
}
}
