% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_adt_dists_fast}
\alias{calc_adt_dists_fast}
\title{Compute per-cell mean ADT distance to its k nearest neighbors from a base assay}
\usage{
calc_adt_dists_fast(adt_data, features, neighbors, exclude_self = TRUE)
}
\arguments{
\item{adt_data}{ADT data matrix (features by cells) to use for distance calculations.}

\item{features}{Optional vector of ADT features; if missing, use all features present.}

\item{neighbors}{The kNN slot.}

\item{exclude_self}{Drop the cell itself from neighbors if present.}
}
\description{
Compute per-cell mean ADT distance to its k nearest neighbors from a base assay
}
\details{
Only returns the mean.
}
