% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_correlation}
\alias{calc_correlation}
\title{Calculate the correlation between each cell's expression and the mean of its neighbors' expression}
\usage{
calc_correlation(seurat_obj, features_adt, cor_method = "spearman")
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{features_adt}{Name of the ADT features to evaluate on (e.g. "CD27.1").}

\item{cor_method}{Correlation method to use (e.g. "pearson", "spearman").}
}
\value{
A data frame with columns: Graph, Feature, Score (correlation value).
}
\description{
Calculate the correlation between each cell's expression and the mean of its neighbors' expression
}
\details{
The Seurat object must have \code{FindNeighbors()} already run at least one time and an assay named "ADT".
}
