% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_distances}
\alias{calc_distances}
\title{This function calculates cluster distances for the given Seurat object.}
\usage{
calc_distances(
  seurat_obj,
  reduction_name,
  criteria = "Within_Max",
  labels_true = "annotated_clusters",
  labels_name
)
}
\arguments{
\item{seurat_obj}{The post-WNN Seurat object.}

\item{reduction_name}{The name of the reduction to use for distances.}

\item{criteria}{One of: Between_Mean, Between_Min, Calinski_Harabasz, Within_Between, Within_Max, Within_Mean, Within_Median.}

\item{labels_true}{The name of the column in the metadata that contains the cluster labels to evaluate.}

\item{labels_name}{A more descriptive name for the labels to use in plotting (optional).}
}
\value{
A data.frame with a row per metric and cluster containing the score.
}
\description{
This function calculates cluster distances for the given Seurat object.
}
\details{
The \code{as.factor()} is needed in case you give a categorical cluster col.
Assumes you're using the embeddings approach.
}
