% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_ext_metrics}
\alias{calc_ext_metrics}
\title{This function calculates external clustering metrics for the given Seurat object.}
\usage{
calc_ext_metrics(
  seurat_obj,
  reduction_name,
  criteria = c("Completeness", "Homogeneity"),
  labels_true = "annotated_clusters",
  labels_pred = "seurat_clusters"
)
}
\arguments{
\item{seurat_obj}{The post-WNN Seurat object.}

\item{reduction_name}{The name of the reduction to use for distances.}

\item{criteria}{One of: ARI, Completeness, Homogeneity}

\item{labels_true}{The name of the column in the metadata that contains the true cluster labels to evaluate.}

\item{labels_pred}{The name of the column in the metadata that contains the predicted cluster labels to evaluate.}
}
\value{
A data.frame with a row per metric containing the combined score.
}
\description{
This function calculates external clustering metrics for the given Seurat object.
}
\details{
The Satija lab used \code{cluster} for their analyses.
The \code{as.factor()} is needed in case you give a categorical cluster col.
sklearn.metrics.cluster has completeness_score and homogeneity_score
Assumes you're using the embeddings approach.
}
