% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calc_int_metrics}
\alias{calc_int_metrics}
\title{Calculate internal clustering metrics}
\usage{
calc_int_metrics(
  seurat_obj,
  reduction_name,
  criteria = "Silhouette",
  labels_true = "annotated_clusters",
  labels_name,
  return_full = FALSE
)
}
\arguments{
\item{seurat_obj}{The post-WNN Seurat object.}

\item{reduction_name}{The name of the reduction to use for distances.}

\item{criteria}{One of: DB, Dunn, Intra_Complete, Silhouette}

\item{labels_true}{The name of the column in the metadata that contains the cluster labels to evaluate.}

\item{labels_name}{A more descriptive name for the labels to use in plotting (optional).}

\item{return_full}{If TRUE, return the full silhouette object instead of just the mean silhouette width.}
}
\value{
A data.frame with a row per metric containing the combined score.
}
\description{
This function calculates various internal clustering metrics using the \code{cluster} package, which provides a comprehensive set of clustering evaluation statistics based on a distance matrix and cluster assignments.
}
\details{
The Satija lab used \code{cluster} for their analyses.
The \code{as.factor()} is needed in case you give a categorical cluster col.
Use the \code{bluster} package for speed if desired e.g.
\code{bluster::approxSilhouette(x = embeddings, clusters = clusters)}
Assumes you're using the embeddings approach.
}
