% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_moran}
\alias{calc_moran}
\title{Calculate Moran's i for a Seurat object}
\usage{
calc_moran(
  seurat_obj,
  features_adt,
  adt_assay = "ADT",
  graph_name,
  row_standardize = TRUE
)
}
\arguments{
\item{seurat_obj}{The Seurat object. Must have \code{FindNeighbors()} already
run.}

\item{features_adt}{Name of the ADT feature to evaluate (e.g. "CD27.1").}

\item{adt_assay}{Name of the assay containing ADT data.}

\item{graph_name}{Name of the neighbor graph slot to use for the weights
matrix (e.g. "RNA.nn", "BCR.nn", "w.nn").}

\item{row_standardize}{Whether to row-standardize the weights matrix (i.e.
make each row sum to 1).}
}
\value{
A single numeric value representing the observed Moran's i.
}
\description{
Calculates the global Moran's i index for an ADT feature given a neighbor
graph.
}
\details{
Row standardization makes sure the resulting score will always be between
-1 and 1.
}
