% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassification.R
\name{calc_nn_frac}
\alias{calc_nn_frac}
\title{This function calculation the proportion of the neighbors that are mutated or unmutated.}
\usage{
calc_nn_frac(seurat_obj, assay = "WNN", plot_title = "", category = "mutated")
}
\arguments{
\item{seurat_obj}{A Seurat object with neighbors calculated.}

\item{assay}{One of: RNA, BCR, WNN}

\item{plot_title}{A string to add to the plot title, e.g. "All Cells" or "Memory B Cells Only".}

\item{category}{One of: Mutated, Unmutated}
}
\value{
A data.frame and a plot.
}
\description{
This function calculation the proportion of the neighbors that are mutated or unmutated.
}
\details{
Assumes that WNN neighbors are stored in "w.nn".
Defines mutated as anything about 0\% SHM.
\code{setNames()} works better than a for loop checking all of the neighbors.
}
