% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{calc_pcts}
\alias{calc_pcts}
\title{Bar plot with stacked percentages.}
\usage{
calc_pcts(
  data,
  meta_group_by = c("sample_id", "Dataset"),
  focus_group,
  order_by
)
}
\arguments{
\item{data}{A data.frame containing the data to be analyzed e.g. the metadata of a Seurat object.}

\item{meta_group_by}{A character vector specifying the columns to group by.}

\item{focus_group}{A character string specifying the column to focus on for percentage calculation.}

\item{order_by}{An optional character string specifying the focus group value to order the results by. If not provided, no ordering is applied.}
}
\value{
A data.frame with the calculated percentages and counts for each group.
}
\description{
This function calculates the percentage of occurrences of a specified focus group within a dataset, grouped by specified metadata columns.
}
\details{
This is not very elegant, but seems to work (dealing with string args is weird).
Make sure to filter your data as needed (e.g. no NA isotypes) beforehand.
Made for cell types and isotypes.
The percentages will be doubles between 0 and 100.
It will fill in missing values (especially needed for ordering if provided).
You could alternatively use the fill option in geom_bar() instead.
}
