% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{cell_type_clusters}
\alias{cell_type_clusters}
\title{This function maps the cell types to the Seurat clusters}
\usage{
cell_type_clusters(
  seurat_obj,
  clusters_col = "seurat_clusters",
  annotations_col
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{clusters_col}{The metadata column with the Seurat clusters.}

\item{annotations_col}{The metadata column with the cell types.}
}
\value{
A data.frame with a row for each cell type.
}
\description{
This function maps the cell types to the Seurat clusters
}
