% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{clean_dimplot}
\alias{clean_dimplot}
\title{Clean theme for DimPlots}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 4.
}
\usage{
clean_dimplot
}
\description{
Clean theme for DimPlots
}
\keyword{datasets}
