% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{clean_dimplot2}
\alias{clean_dimplot2}
\title{Clean theme for DimPlots}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 5.
}
\usage{
clean_dimplot2
}
\description{
Clean theme for DimPlots
}
\details{
Same as clean_dimplot but with a square aspect ratio.
}
\keyword{datasets}
