% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{clean_umap}
\alias{clean_umap}
\title{Clean theme for UMAP plots}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 5.
}
\usage{
clean_umap
}
\description{
Clean theme for UMAP plots
}
\keyword{datasets}
