% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{extract_wnn_vars}
\alias{extract_wnn_vars}
\title{This function gives a summary of the post-WNN object.}
\usage{
extract_wnn_vars(
  seurat_obj,
  gex_pca = "rpca",
  other_pca = "bpca",
  other_type = "BCR"
)
}
\arguments{
\item{seurat_obj}{The post-WNN Seurat object}

\item{gex_pca}{The name of the GEX PCA reduction.}

\item{other_pca}{The name of the BCR/ADT/etc. PCA reduction.}

\item{other_type}{The second assay. Defaults to "BCR".}
}
\value{
A text message.
}
\description{
This function gives a summary of the post-WNN object.
}
\details{
Assumes that embeddings were used (for now) and that the object has RNA.
}
