% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airr.R
\name{factor_family_info}
\alias{factor_family_info}
\title{Convert family and gene information to factors with proper ordering.}
\usage{
factor_family_info(combined_airr)
}
\arguments{
\item{combined_airr}{An AIRR-formatted data.frame. This function ensures that gene
families and genes are ordered correctly (e.g., IGHV1, IGHV2, IGHV10 instead
of IGHV1, IGHV10, IGHV2).}
}
\value{
A data.frame with up to six columns converted to properly ordered factors.
}
\description{
Converts the gene family and gene columns added by add_family_info() to
properly ordered factors using numeric sorting.
}
\details{
For after add_family_info() has been run
}
