% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{find_k_clusters}
\alias{find_k_clusters}
\title{Find the right clustering resolution to obtain desired number of clusters}
\usage{
find_k_clusters(seurat_obj, graph_name = "RNA_snn", desired_k)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{graph_name}{The name of the graph to use for clustering. Default is "RNA_snn".}

\item{desired_k}{The desired number of clusters.}
}
\value{
The Seurat object with clusters at the resolution that matches desired_k.
}
\description{
This function iteratively tests clustering resolutions in Seurat to find the resolution
that yields the desired number of clusters. It uses the specified graph and returns the
Seurat object with clusters if successful, or stops if the desired number is exceeded or not found.
}
