% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{find_n_clusters}
\alias{find_n_clusters}
\title{Find the right clustering resolution to obtain the desired number of clusters}
\usage{
find_n_clusters(
  seurat_obj,
  graph_name = "RNA_snn",
  desired_n,
  res_range = seq(0.1, 2, by = 0.1)
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{graph_name}{The name of the graph to use for clustering.}

\item{desired_n}{The desired number of clusters.}

\item{res_range}{The range of resolutions to test.}
}
\value{
The Seurat object with clusters at the resolution that matches \code{desired_n}.
}
\description{
This function iteratively tests clustering resolutions in Seurat to find the resolution
that yields the desired number of clusters. It uses the specified graph and returns the
Seurat object with clusters if successful, or stops if the desired number is exceeded or not found.
}
\details{
This would typically be used after \code{\link[=seurat_pipeline]{seurat_pipeline()}}.
}
