% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{gen_dot_title}
\alias{gen_dot_title}
\title{Generates a title for the DotPlot}
\usage{
gen_dot_title(plot_title = "", marker_sources)
}
\arguments{
\item{plot_title}{Dataset description}

\item{marker_sources}{The list of marker sources}
}
\value{
A string with sources comma-separated and in parentheses
}
\description{
Generates a title for the DotPlot
}
