% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{get_cell_types}
\alias{get_cell_types}
\title{Pulls the info needed for \code{add_info_bar()} to a \code{DotPlot}}
\usage{
get_cell_types(markers_df)
}
\arguments{
\item{markers_df}{The markers data.frame filtered to match your input features}
}
\value{
A data.frame with Cell_Type_Full and features.plot columns
}
\description{
Pulls the info needed for \code{add_info_bar()} to a \code{DotPlot}
}
