% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex_airr.R
\name{gex_add_airr}
\alias{gex_add_airr}
\title{Add AIRR information to a Seurat object}
\usage{
gex_add_airr(
  seurat_obj,
  airr_type = "BCR",
  combined_airr,
  new_cols,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{The Seurat object containing GEX data.}

\item{airr_type}{Type of immune receptor data. Currently supports "BCR".}

\item{combined_airr}{BCR AIRR formatted data frame with heavy and light chains.}

\item{new_cols}{Vector of column names to select from the AIRR data.}

\item{verbose}{Logical indicating whether or not to print integration summary information.}
}
\value{
The Seurat object with AIRR columns added to the metadata, including
Has_BCR, isotype information, mutation frequencies, and pairing status.
}
\description{
This function integrates adaptive immune receptor repertoire (AIRR) data with
gene expression (GEX) data in a Seurat object. Currently built to integrate
BCR data, including both heavy and light chain information.
}
\details{
This would typically be used after \code{\link[=seurat_pipeline]{seurat_pipeline()}} and before \code{\link[=concatenate_gex_bcr]{concatenate_gex_bcr()}}.
Right now this is just built to integrate in BCR data and assumes that the BCR data includes light chains.
Assumes that \code{seurat_obj} contains \code{cell_id} and \code{annotated_clusters_simpler}.
Collapsed light chains are alphabetized.
There are columns for IGK and IGL instead of just "light" since a cell can have both.
Adds mutation frequency bins and isotype staging information for BCR data.
}
