% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex_airr.R
\name{infer_k_param}
\alias{infer_k_param}
\title{Infer k_param from existing RNA neighbors, or return a default.}
\usage{
infer_k_param(seurat_obj, default = 20, verbose = TRUE)
}
\arguments{
\item{seurat_obj}{A Seurat object containing a neighbors slot.}

\item{default}{The default k_param to return if not found in the neighbors slot.}

\item{verbose}{Logical indicating whether or not to print a message about the source of the
inferred k_param.}
}
\value{
An integer representing the k_param used for neighbor finding.
}
\description{
Infers the k_param used for neighbor finding from the existing RNA neighbors slot, or
returns a default if not found.
}
