% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex_airr.R
\name{infer_num_dims}
\alias{infer_num_dims}
\title{Infer num_dims (PCA dims used for neighbor finding) from existing RNA neighbors.}
\usage{
infer_num_dims(seurat_obj, default = 20, verbose = TRUE)
}
\arguments{
\item{seurat_obj}{A Seurat object containing a neighbors slot.}

\item{default}{The default number of dimensions to return if not found in the neighbors slot.}

\item{verbose}{Logical indicating whether or not to print a message about the source of the inferred num_dims.}
}
\value{
An integer representing the number of dimensions used for neighbor finding.
}
\description{
Infers the number of dimensions used for neighbor finding from the existing RNA neighbors slot, or returns a default if not found.
}
