% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{labels_rotate_x}
\alias{labels_rotate_x}
\title{Rotate x-axis labels 45 degrees}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 1.
}
\usage{
labels_rotate_x
}
\description{
Rotate x-axis labels 45 degrees
}
\keyword{datasets}
