% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{labels_standard}
\alias{labels_standard}
\title{Standard label sizes for athanor plots}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 7.
}
\usage{
labels_standard
}
\description{
Standard label sizes for athanor plots
}
\keyword{datasets}
