% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{labels_standard_vln}
\alias{labels_standard_vln}
\title{Standard label sizes for violin plots}
\usage{
labels_standard_vln
}
\description{
Standard label sizes for violin plots
}
\details{
Removes "Identity".
}
