% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{labels_standard_vln_rotate}
\alias{labels_standard_vln_rotate}
\title{Violin plot labels with horizontal x-axis text}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 7.
}
\usage{
labels_standard_vln_rotate
}
\description{
Violin plot labels with horizontal x-axis text
}
\keyword{datasets}
