% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{layout_doublets}
\alias{layout_doublets}
\title{This function plots an overview of a doublet identification method}
\usage{
layout_doublets(
  seurat_obj,
  tissue_type,
  clrs_specific,
  use_hues = FALSE,
  group_col = "seurat_clusters",
  group_label = NULL,
  doublet_col = "scDblFinder.class",
  doublet_package = "scDblFinder",
  details = NULL
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{tissue_type}{Blood, Skin.}

\item{clrs_specific}{The specific color palette (should be named).}

\item{use_hues}{Use the iwanthue hues instead of the default ggplot colors. Doesn't let you set any other settings.}

\item{group_col}{The column to group by.}

\item{group_label}{The label for the grouping variable to use in the plot titles and axis labels. If NULL, it will be determined based on the group_col name.}

\item{doublet_col}{The column containing the doublets information}

\item{doublet_package}{The doublet method being used.}

\item{details}{The optional subtitle.}
}
\value{
A grid of four plots with UMAPs in the left column and bar plots in the right column.
}
\description{
This function plots an overview of a doublet identification method
}
\details{
It assumes that named_colors$doublet has been defined.
Depends on other plots.
The doublets will be plotted "on top" for the first UMAP.
}
