% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{named_colors}
\alias{named_colors}
\title{Named color palettes for athanor plots}
\format{
An object of class \code{list} of length 25.
}
\usage{
named_colors
}
\description{
A list of named character vectors mapping categorical values to hexadecimal colors, typically used as the \code{clrs_specific} argument in plotting functions. These color mappings are used to ensure consistent and meaningful coloring across different plots in the athanor package, especially for categories that are commonly visualized (e.g. cell types, isotypes, SHM frequencies).
}
\details{
The \code{named_colors} list contains color mappings for various categories relevant to BCR and single-cell analysis, including:
\itemize{
\item \code{c_call}: Colors for IG constant region calls (e.g. IGHA, IGHD, etc.)
\item \code{cdr3}: Colors for CDR3 lengths from 4 to 41
\item \code{cell_types_celltypist}, \code{cell_types_simpler}: Colors for cell type annotations from CellTypist and a simpler scheme
\item \code{datatype}: Colors for different data types (e.g. ADT, BCR, GEX, etc.)
\item \code{doublet}: Colors for singlets vs doublets in doublet detection
\item \code{embeddings}: Colors for different embedding methods (e.g. antiberta2, balm-paired, etc.)
\item \code{isotype_stage}: Colors for isotype switching stages (i.e. Unswitched, Switched)
\item \code{isotype}: Colors for immunoglobulin isotypes (e.g. IgA, IgD, etc.)
\item \code{light}: Colors for light chain types (i.e. IGK, IGL, IGK/IGL)
\item \code{mu_freq_bins}, \code{mu_freq_bins_fewer}, \code{mu_freq_bins_binary}: Colors for binned mutation frequencies
\item \code{mu_freq_iso}: Colors for combinations of mutation frequency bins and isotype switching stages
\item \code{v_call_family}, \code{d_call_family}, \code{j_call_family}: Colors for V, D, J gene families respectively
\item \code{weights}: Colors for WNN weights from 0 to 1
}

When possible, colors are chosen to be colorblind-friendly.
}
\keyword{datasets}
