% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{plot_anno}
\alias{plot_anno}
\title{Patchwork annotation with Roman numeral panel labels}
\format{
An object of class \code{plot_annotation} of length 8.
}
\usage{
plot_anno
}
\description{
Patchwork annotation with Roman numeral panel labels
}
\keyword{datasets}
