% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_color_scale}
\alias{plot_color_scale}
\title{Generates a proper color scale for a Seurat DotPlot (with white at zero).}
\usage{
plot_color_scale(
  plot,
  palette = rev(pals::brewer.rdbu(n = 7)),
  val_col = "avg.exp.scaled",
  fill_by = "color"
)
}
\arguments{
\item{plot}{The generated Seurat DotPlot or ggplot.}

\item{palette}{A palette of colors to go off of.}

\item{val_col}{The column in the plot data that contains the values to be plotted (e.g. "avg.exp.scaled").}

\item{fill_by}{One of "color" or "fill".}
}
\value{
A Seurat dot plot or ggplot with an updated color scale.
}
\description{
Generates a proper color scale for a Seurat DotPlot (with white at zero).
}
\details{
Seurat's \code{col} option frequently is misleading with where the zeroes fall.
I don't want to rescale the expression.
You could also just do something like \code{scale_color_gradient2(low = "#2166AC", mid = "white", high = "#B2182B")}
Can also be used to generated a color scale for a general ggplot2.
Use the function via a pipe right after the function call.
}
