% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dimplot}
\alias{plot_dimplot}
\title{Plot a Seurat UMAP using \code{DimPlot}}
\usage{
plot_dimplot(
  seurat_obj,
  data_source = "",
  clrs_specific,
  use_hues = FALSE,
  pt_size = 0.2,
  assay,
  reduc = "umap",
  plot_label = TRUE,
  label_box = TRUE,
  label_size = 3,
  annotated = FALSE,
  specific_clusters,
  meta_col = "seurat_clusters",
  include_legend = TRUE,
  legend_label,
  factor_idents = TRUE,
  order = FALSE,
  details,
  ...
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{data_source}{The dataset of origin.}

\item{clrs_specific}{Specific colors for plotting (make sure it has names).}

\item{use_hues}{Use the iwanthue hues instead of the default ggplot colors. Doesn't let you set any other settings.}

\item{pt_size}{The point size.}

\item{assay}{The data type e.g. ADT, GEX, BCR, WNN...}

\item{reduc}{The reduction to use for plotting e.g. wnn.umap}

\item{plot_label}{Add labels to the plot (or not).}

\item{label_box}{Whether or not to give the labels a background.}

\item{label_size}{The size of the plot labels.}

\item{annotated}{If the cell types have been identified.}

\item{specific_clusters}{Can overlay clusters of interest e.g. B cell or by #. Overrides the annotated option.}

\item{meta_col}{Which column in the object metadata to color by.}

\item{include_legend}{Include the legend or not.}

\item{legend_label}{The label for the legend.}

\item{factor_idents}{Whether or not to factorize the idents (for proper ordering of the colors). This can mess up the order you want, so be careful.}

\item{order}{Plot cells on top or not.}

\item{details}{A custom subtitle.}

\item{...}{Any other Seurat parameters.}
}
\value{
A Seurat UMAPPlot.
}
\description{
This function generates a UMAP plot from a Seurat object using \code{DimPlot} with various customizable options for coloring, labeling, and grouping the data.
}
