% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dot_airr}
\alias{plot_dot_airr}
\title{Add AIRR (and other) info along the right side of an existing Seurat DotPlot.}
\usage{
plot_dot_airr(
  plot,
  seurat_obj,
  row_identity = "seurat_clusters",
  facet_col,
  info_to_add = c("cluster_size")
)
}
\arguments{
\item{plot}{The generated Seurat DotPlot.}

\item{seurat_obj}{The Seurat object containing the data.}

\item{row_identity}{The y axis identities.}

\item{facet_col}{The column to facet by e.g. "Cell_Type_Full".}

\item{info_to_add}{Vector of information types to add. The options are: \code{cluster_size}, \code{mean_mu_freq}, \code{percent_BCR}, \code{percent_TCR}, and \code{TRUST4}.}
}
\value{
A Seurat DotPlot with annotations along the right side.
}
\description{
This function enhances dot plots by adding adaptive immune receptor repertoire (AIRR)
information and other metadata along the right side for easy comparison. Currently
supports cluster size, mean mutation frequency, and BCR/TCR percentages.
}
\examples{
\dontrun{
p <- DotPlot(seurat_obj, features = genes)
plot_dot_airr(p, seurat_obj, info_to_add = c("cluster_size", "mean_mu_freq"))
}
}
