% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_doublets}
\alias{plot_doublets}
\title{Plot an overview of a doublet identification method}
\usage{
plot_doublets(
  seurat_obj,
  tissue_type,
  clrs_specific,
  use_hues = FALSE,
  group_col = "seurat_clusters",
  group_label = NULL,
  doublet_col = "scDblFinder.class",
  doublet_package = "scDblFinder",
  details = NULL
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{tissue_type}{Blood, Skin.}

\item{clrs_specific}{The specific color palette (should be named).}

\item{use_hues}{Use the \code{iwanthue} hues instead of the default ggplot colors. Doesn't let you set any other settings.}

\item{group_col}{The column to group by.}

\item{group_label}{The label for the grouping variable to use in the plot titles and axis labels. If NULL, it will be determined based on the group_col name.}

\item{doublet_col}{The column containing the doublets information}

\item{doublet_package}{The doublet method being used.}

\item{details}{The optional subtitle.}
}
\value{
A grid of four plots with UMAPs in the left column and bar plots in the right column.
}
\description{
This function creates a grid of four plots to visualize the results of a doublet identification method.
The left column contains UMAP plots colored by doublet/singlet status and by clusters/annotations, while the right column contains bar plots showing the counts and percentages of doublets across clusters or annotations.
}
\details{
It assumes that \code{named_colors$doublet} has been defined.
Depends on other plots.
The doublets will be plotted "on top" for the first UMAP.
}
