% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airr.R
\name{plot_immune_overlay}
\alias{plot_immune_overlay}
\title{Plot UMAPs with AIRR overlays}
\usage{
plot_immune_overlay(
  seurat_obj,
  tissue_type,
  airr_type,
  clrs_specific,
  plot_by = "all",
  barcode_col = "cell_id",
  plot_label = FALSE,
  ncol_sample = 6
)
}
\arguments{
\item{seurat_obj}{The Seurat object (must contain Has_BCR/Has_TCR cols)}

\item{tissue_type}{The tissue type of interest.}

\item{airr_type}{BCR or TCR}

\item{clrs_specific}{The specific color palette (should be named).}

\item{plot_by}{The grouping method: all samples, by dataset, by sample, by isotype.}

\item{barcode_col}{The barcode column name: barcode, cell_id, or Cell_ID_Unique.}

\item{plot_label}{Whether or not to include the labels.}

\item{ncol_sample}{The number of columns for sample-wise plots.}
}
\value{
A Seurat UMAP plot with AIRR cells highlighted by the specified grouping.
}
\description{
This function plots the annotated UMAPs alongside the BCR/TCR overlays.
}
\details{
lightgray is Seurat's default background color and cells_total has to be a list for the labels to work later.
This could probably be replaced with \code{plot_umap()}.
}
