% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot the internal or external clustering metrics}
\usage{
plot_metrics(
  metrics,
  plot_title = "",
  best_score = "higher",
  type = "Internal",
  y_axis = "Labeling",
  round_to = 2,
  details = ""
)
}
\arguments{
\item{metrics}{Data frame of metrics to plot, with columns: Embedding, Reduction, Labeling, Metric, Score.}

\item{plot_title}{Title to use for the plot.}

\item{best_score}{One of "higher" or "lower" to indicate whether higher or lower scores are better for the metrics being plotted. This is used to determine which scores to outline in the plot.}

\item{type}{One of "Internal" or "External" to indicate the type of metrics being plotted, used for the plot title.}

\item{y_axis}{The name of the column in the metrics data frame to use for the y-axis (e.g. "Labeling").}

\item{round_to}{Number of decimal places to round the score labels to in the plot.}

\item{details}{Additional details to include in the plot title (optional).}
}
\value{
A ggplot showing the metrics across embeddings and reductions, with the best scores outlined.
}
\description{
This function creates a ggplot visualization of the internal or external clustering metrics across different embeddings and reductions.
It uses a color scale to represent the metric scores and outlines the best scores for each label and metric combination, as well as the best scores across reductions.
The plot is faceted by metric and reduction for easy comparison.
}
\details{
Make sure to check that the best score is consistent across your plotting metrics.
}
