% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mws}
\alias{plot_mws}
\title{Plot a box plot of modality weights per cell type}
\usage{
plot_mws(
  seurat_obj,
  second_assay = "BCR",
  clrs_specific = named_colors$mu_freq_bins,
  split_by = "mu_freq_bins",
  facet_by = "annotated_clusters",
  y_axis_label = "SHM Frequency Bins",
  details = NULL
)
}
\arguments{
\item{seurat_obj}{The post-WNN Seurat object.}

\item{second_assay}{List of other assays run through WNN in order.}

\item{clrs_specific}{The specific color palette (should be named).}

\item{split_by}{A meta.data column to split the box plots up by.}

\item{facet_by}{A meta.data column to facet the box plots by.}

\item{y_axis_label}{Label for the y-axis.}

\item{details}{An optional custom subtitle.}
}
\value{
A ggplot with the distribution of weights
}
\description{
This function creates a box plot to visualize the distribution of modality weights (e.g. RNA vs. BCR) across different cell types or clusters in a post-WNN Seurat object.
It should be run after \code{\link[=run_wnn]{run_wnn()}}.
}
\details{
Assumes "annotated_clusters" is a column.
}
