% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_overview_comps}
\alias{plot_overview_comps}
\title{Plots several UMAPs side by side for a Seurat object.}
\usage{
plot_overview_comps(
  seurat_objs,
  data_source = "",
  pt_size = 0.1,
  second_assay = "BCR",
  assay_name,
  reduction = "wnn.umap",
  comparisons = c("annotated_clusters", "v_call_family", "light_chains", "isotype",
    "mu_freq")
)
}
\arguments{
\item{seurat_objs}{List of WNN objects with different embedding types.}

\item{data_source}{Dataset description.}

\item{pt_size}{The size of the points in the UMAP.}

\item{second_assay}{The second assay to use in the title if plotting a WNN reduction.}

\item{assay_name}{The name of the assay to use in the title. By default, it will be set based on the reduction (e.g. "GEX" for "rna.umap", "BCR" for "bcr.umap", and "GEX & BCR" for "wnn.umap").}

\item{reduction}{Which reduction to plot ("rna.umap", "bcr.umap", or "wnn.umap").}

\item{comparisons}{The labelling of the plots.}
}
\value{
patchwork object with overview plots
}
\description{
Plots several UMAPs side by side for a Seurat object.
}
\details{
The names of the seurat_objs list correspond with embedding_types.
Assumes that CellTypist is the annotation approach being used.
data_source is set to empty to save on space
}
