% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pcts}
\alias{plot_pcts}
\title{Plot percentages in a stacked bar plot}
\usage{
plot_pcts(
  pcts,
  tissue_type,
  clrs_specific,
  plot_type = "All",
  plot_value = "Cell Type",
  x_axis = "sample_id",
  x_axis_label = "Sample",
  fill_type = "annotated_clusters",
  fill_label = fill_type,
  perc_min = 3,
  label_size = 3,
  label_fill = FALSE,
  include_counts = TRUE,
  drop_zeroes = TRUE,
  reverse_order = FALSE,
  total_order = FALSE,
  details = NULL
)
}
\arguments{
\item{pcts}{The output of \code{calc_pcts()}.}

\item{tissue_type}{The type of tissue being plotted e.g. Blood or Skin.}

\item{clrs_specific}{A specific (must have names) color palette.}

\item{plot_type}{One of \code{All}, \code{Binary}.}

\item{plot_value}{What is being plotted.}

\item{x_axis}{What to put along the x axis.}

\item{x_axis_label}{The label for the x axis.}

\item{fill_type}{What to group the bar plot by.}

\item{fill_label}{The description of what you're filling by.}

\item{perc_min}{The minimum percentage to show in the plot.}

\item{label_size}{The size of the percentage labels.}

\item{label_fill}{Add a white background for clarity.}

\item{include_counts}{Plot the counts on top/bottom.}

\item{drop_zeroes}{Remove percentages of zeroes.}

\item{reverse_order}{Change the fill order for a stacked plot.}

\item{total_order}{Rearrange x axis in descending order by totals instead of alphabetically.}

\item{details}{The optional subtitle.}
}
\value{
A stacked ggplot bar plot
}
\description{
This function plots a stacked bar plot of percentages calculated using \code{\link[=calc_pcts]{calc_pcts()}} with percentages
labeled and total counts on top.
}
\details{
Give the percentages already as percents (* 100 in the calculations).
Make sure \code{pcts} includes Dataset if you want to split by dataset.
Note that the percentages may seem inaccurate because of the accuracy.
There was a big issue with the combo of \code{geom_text()} & \code{facet_grid()}.
This assumes that you want to show the counts for binary plots.
}
