% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_umap_condition}
\alias{plot_umap_condition}
\title{Plot a specific condition on a Seurat UMAP}
\usage{
plot_umap_condition(
  seurat_obj,
  tissue_type,
  clrs_specific,
  condition_name,
  operator,
  condition_val,
  color_by = "value",
  plot_type = "general",
  label_plot = TRUE,
  include_subtitle = TRUE,
  include_legend = FALSE
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{tissue_type}{The tissue type e.g. Blood, Skin.}

\item{clrs_specific}{The specific color palette (should be named).}

\item{condition_name}{The column in the object that contains the condition of interest e.g. "annotated_clusters", "mu_freq", "isotype", etc.}

\item{operator}{<, >, ==}

\item{condition_val}{The value to compare the condition to. If color_by is "name", this should be a name in the condition_name column. If color_by is "value", this should be a value in the condition_name column.}

\item{color_by}{name, value}

\item{plot_type}{general, overlay (BCR/TCR)}

\item{label_plot}{Put labels on the plot (or not).}

\item{include_subtitle}{Include a subtitle (or not).}

\item{include_legend}{Include a legend (or not).}
}
\value{
A Seurat UMAPPlot.
}
\description{
This function generates a UMAP plot from a Seurat object using \code{DimPlot} with various customizable options for coloring, labeling, and grouping the data based on a specific condition of interest.
}
\details{
Based on: https://github.com/satijalab/seurat/issues/1053
Can be used for plotting QC metrics, isolating specific cell types, overlaying AIRR data, overlaying B cell isotypes, etc.
}
