% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_vln_feat}
\alias{plot_vln_feat}
\title{Plot a Seurat \code{VlnPlot} and a \code{FeaturePlot} side by side for the same marker}
\usage{
plot_vln_feat(
  seurat_obj,
  clrs_specific,
  feature,
  reduc = "umap",
  meta_col = NULL,
  rotate = FALSE,
  ...
)
}
\arguments{
\item{seurat_obj}{The Seurat object with GEX data.}

\item{clrs_specific}{The specific color palette (should be named).}

\item{feature}{The feature of interest.}

\item{reduc}{The reduction to use for plotting e.g. "bpca" or wnn.umap".}

\item{meta_col}{What to group by (uses the Idents by default).}

\item{rotate}{Rotate the labels or not.}
}
\value{
Two patchworked Seurat plots.
}
\description{
This function generates a side-by-side visualization of a Seurat \code{VlnPlot} and a \code{FeaturePlot} for a specified marker, allowing for a comprehensive comparison of expression levels across different groups and spatial distribution on the UMAP.
The \code{VlnPlot} will display the distribution of expression levels across specified groups, while the \code{FeaturePlot} will show the spatial localization of the marker on the UMAP, with higher expressing cells highlighted on top for better visibility.
}
\details{
Will put the highest expressing cells on top for the latter.
}
