% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{plot_wnn_testing}
\alias{plot_wnn_testing}
\title{Plot the results of manual WNN testing}
\usage{
plot_wnn_testing(manual_wnn_test, sim_var, other_vars, count_range)
}
\arguments{
\item{manual_wnn_test}{A data frame containing the results of the manual WNN testing, with columns "Count" and "Passed".}

\item{sim_var}{The variable that was varied in the manual WNN testing (e.g. "Genes", "Cells", "Dimensions", "GEX PCs", or "BCR PCs").}

\item{other_vars}{A named list of the other variables that were held constant in the manual WNN testing, with names corresponding to the variable names (e.g. "Genes", "Cells", "Dimensions", "GEX PCs", or "BCR PCs") and values corresponding to the constant values used in the testing.}

\item{count_range}{A numeric vector specifying the range of counts that were tested in the manual WNN testing, used for setting the x-axis breaks in the plot.}
}
\value{
A ggplot object showing the step plot of the WNN testing results, with the x-axis representing the count of the varied variable and the y-axis representing whether Seurat's WNN computation passed (1), failed (0), or gave a warning (0.5). The plot includes a title indicating the variable that was varied and a subtitle listing the other variables that were held constant.
}
\description{
This function creates a step plot to visualize the results of manual WNN testing, showing how the success of Seurat's WNN computation changes as the count of a specific variable is varied while keeping other variables constant.
}
