% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{plot_wnn_umaps}
\alias{plot_wnn_umaps}
\title{This function gives a summary of the post-WNN object.}
\usage{
plot_wnn_umaps(
  seurat_obj,
  data_source = "Manual",
  airr_type = "BCR",
  airr_processing = "Embeddings",
  reducs_list = c("rna.umap", "bcr.umap", "wnn.umap"),
  clusters_col = "seurat_clusters",
  plot_label = TRUE,
  pt_size = 0.8,
  clrs_specific
)
}
\arguments{
\item{seurat_obj}{The post-WNN Seurat object.}

\item{data_source}{The source of the data for the plot title.}

\item{airr_type}{The type of AIRR data.}

\item{airr_processing}{The type of AIRR processing; one of \verb{c("Embeddings", "Features)}}

\item{reducs_list}{Should be in order (GEX, AIRR, WNN).}

\item{clusters_col}{The metadata column to color the UMAPs by.}

\item{plot_label}{Whether or not to plot the metadata labels on the UMAPs.}

\item{pt_size}{The point size for the UMAPs.}

\item{clrs_specific}{A specific (must have names) color palette for the clusters. If not provided, the default Seurat colors will be used.}
}
\value{
A combined plot of the GEX, BCR, and WNN UMAPs colored by the specified metadata column.
}
\description{
This function gives a summary of the post-WNN object.
}
\details{
Should be able to plot ADT instead of BCR too.
Only plots one clustering.
}
