% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_dt}
\alias{print_dt}
\title{Display a sortable, scrollable \code{DataTable} table}
\usage{
print_dt(table, dt_width = "800px")
}
\arguments{
\item{table}{The input data frame to be printed.}

\item{dt_width}{The width of the output table.}
}
\value{
A DataTable table.
}
\description{
Display a sortable, scrollable \code{DataTable} table
}
