% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_kable}
\alias{print_kable}
\title{Display a nicely formatted table}
\usage{
print_kable(table, kable_height = "500px", kable_width = "100\%")
}
\arguments{
\item{table}{The input data frame to be printed}

\item{kable_height}{The height of the output table (you can set it to NULL to
display the full table without scrolling). Default is "500px".}

\item{kable_width}{The width of the output table. Default is "100\%".}
}
\value{
A formatted table.
}
\description{
This function creates a formatted table with scrolling capability for use in
R Markdown documents. It applies striped styling and makes the table scrollable
within specified dimensions.
}
