% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{print_metadata_summary}
\alias{print_metadata_summary}
\title{Give an overview of the data in terms of files, samples and datasets}
\usage{
print_metadata_summary(meta_file)
}
\arguments{
\item{meta_file}{The analyst-created metadata csv file containing disease,
sample, and subject information.}
}
\value{
A text description of the provided metadata file.
}
\description{
This function provides a summary of metadata files, printing information about the number of subjects, samples, and total files for each disease in the dataset.
Can handle metadata files representing multiple diseases and datasets.
}
