% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airr.R
\name{process_airrflow}
\alias{process_airrflow}
\title{This function reads airrflow's repertoire files, processes them to add
subject and sample information, computes CDR3 amino acid properties, and adds
gene family information.}
\usage{
process_airrflow(dataset_path, version_airrflow)
}
\arguments{
\item{dataset_path}{The path to the dataset directory.}

\item{version_airrflow}{The airrflow version (as a string).}
}
\value{
A processed AIRR-formatted data.frame with several columns added.
}
\description{
This function reads airrflow's repertoire files, processes them to add
subject and sample information, computes CDR3 amino acid properties, and adds
gene family information.
}
\details{
Only written for BCR data right now.
v4.0 has \code{clone_size_count} and \code{clone_size_freq} columns.
v4.3.1 has \code{duplicate_count} and \code{light_only_cell} columns.
There can still be some \code{NA} \code{c_call}s.
}
