% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{reformat_vdj_barcode_sample}
\alias{reformat_vdj_barcode_sample}
\title{Reformat VDJ barcodes by adding sample names and removing suffixes}
\usage{
reformat_vdj_barcode_sample(data, sample_name, barcode_col = "barcode")
}
\arguments{
\item{data}{The input dataset which contains a barcode column.}

\item{sample_name}{The sample name to add to make the barcode unique.}

\item{barcode_col}{The column in the data that contains the barcodes.}
}
\value{
A character vector of reformatted barcodes.
}
\description{
This function adds the sample name to the beginning of barcodes and removes
the "-1" suffixes from the ends to create unique cell identifiers.
}
