% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{run_wnn}
\alias{run_wnn}
\title{Run Weighted Nearest Neighbors (WNN) analysis on combined GEX and BCR data}
\usage{
run_wnn(
  seurat_obj,
  embeddings,
  embedding_type,
  pc_gex = 20,
  pc_bcr = 20,
  k_param = 20,
  k_main = 20,
  cluster = FALSE,
  cluster_res = list(GEX = 1, BCR = 1, WNN = 1),
  modality_weights = NULL,
  show_output = FALSE
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing GEX data (at the least).}

\item{embeddings}{Matrix of BCR embeddings (genes x cells).}

\item{embedding_type}{The embeddings method.}

\item{pc_gex}{The number of PCs for the GEX assay.}

\item{pc_bcr}{The number of PCs for the BCR assay.}

\item{k_param}{The number of neighbors to use for each modality. Can be a single value or a vector of values to test.}

\item{k_main}{The main number of neighbors to use for the final WNN UMAP and clustering.}

\item{cluster}{Whether or not to perform clustering.}

\item{cluster_res}{Named list of clustering resolutions for GEX, BCR, and WNN.}

\item{modality_weights}{Named vector of modality weights. If NULL, Seurat will calculate automatically.}

\item{show_output}{Whether or not to show verbose output from Seurat functions.}
}
\value{
A Seurat object with WNN run.
}
\description{
This function takes a Seurat object containing gene expression (GEX) data and a matrix of BCR embeddings, and performs WNN analysis to integrate the two modalities.
It processes each assay separately (normalization, variable feature selection, scaling, PCA, neighbor finding, UMAP), then finds multimodal neighbors and runs clustering if specified.
The function also adds metadata about modality weights and run information to the resulting Seurat object.
}
\details{
\itemize{
\item Currently only works for the embeddings approach and BCR data.
\item The GEX object must have a \code{cell_id} metadata column.
\item If I end up combining multiple embeddings into one object, then I should use
something like \code{bcr_assay_name <- "BCR" # paste0("BCR_", embedding_type)}
\item The neighbors for the assays can be saved in both the \code{graphs} slot (\code{compute.SNN} for clustering) and the \code{neighbors} slot (\code{return.neighbors} for distance calculations later)
The \code{compute.SNN} option constructs a shared nearest neighbor graph using Jaccard index.
Perhaps the GEX and BCR sections should be run if \code{modality_weights} is provided.
}
}
