% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_wnn_sims}
\alias{run_wnn_sims}
\title{Run manual WNN simulations by varying a specific variable while keeping others constant}
\usage{
run_wnn_sims(count_range, sim_var, other_vars, show_progress = FALSE)
}
\arguments{
\item{count_range}{A vector of values for the variable to be varied in the simulations.}

\item{sim_var}{The name of the variable being varied (e.g. "Genes", "Cells", "Dimensions", "GEX PCs", "BCR PCs").}

\item{other_vars}{A list of other variables and their constant values to be used in the simulations.}

\item{show_progress}{Whether to display a progress bar during the simulations.}
}
\value{
A data frame with columns "Count" and "Passed", where "Count" corresponds to the values in \code{count_range} and "Passed" indicates whether the simulation run passed without errors (1), had warnings (0.5), or had errors (0).
}
\description{
This function runs a series of WNN simulations by varying a specified variable (e.g. number of genes, cells, dimensions, GEX PCs, BCR PCs) while keeping other parameters constant. It captures whether each simulation run passes without errors or warnings and returns a data frame summarizing the results.
}
\details{
It uses both \code{\link[=sim_airr_manual]{sim_airr_manual()}} and \code{\link[=sim_gex_manual]{sim_gex_manual()}} to generate the necessary data for the WNN simulations. The function also includes error handling to capture any issues that arise during the simulation runs and provides an option to display a progress bar.
}
