% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{sim_bcr_manual}
\alias{sim_bcr_manual}
\title{Simulate an assay made from BCR embeddings independent of any GEX object}
\usage{
sim_bcr_manual(num_cells, num_dims, separator)
}
\arguments{
\item{num_cells}{Number of cells to simulate.}

\item{num_dims}{Number of embedding dimensions to simulate.}

\item{separator}{Separator for cell and dimension names.}
}
\value{
A Seurat Assay
}
\description{
This function simulates an assay of BCR embeddings by generating a matrix of random values between -0.6 and 0.6, with a specified number of cells and embedding dimensions.
}
\details{
The choice of value ranges was based off of real data, including having 9 decimal points.
}
