% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{sim_gex_manual}
\alias{sim_gex_manual}
\title{Manual simulation of GEX data.}
\usage{
sim_gex_manual(num_genes = 1000, num_cells = 2000, separator = "-")
}
\arguments{
\item{num_genes}{Number of genes to simulate.}

\item{num_cells}{Number of cells to simulate.}

\item{separator}{Separator for gene and cell names.}
}
\value{
A Seurat object
}
\description{
Manual simulation of GEX data.
}
\details{
Partially based off of \href{http://gateway.training.ncgr.org/single-cell-workshop/seurat.html}{Single Cell workshop: Chapter 6}
}
