% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{sim_gex_splatter}
\alias{sim_gex_splatter}
\title{Simulate gene expression data using Splatter}
\usage{
sim_gex_splatter(
  num_genes = 1000,
  num_cells = 2000,
  splatter_groups = 1,
  splatter_method = "single",
  seed = 42,
  verbose = FALSE
)
}
\arguments{
\item{num_genes}{Number of genes to simulate. Defaults to 1000.}

\item{num_cells}{Number of cells to simulate. Defaults to 2000.}

\item{splatter_groups}{Group probabilities for Splatter simulation. Defaults to 1.}

\item{splatter_method}{Splatter simulation method. Defaults to "single".}

\item{seed}{Random seed for reproducible simulations. Defaults to 42.}

\item{verbose}{Whether or not to print verbose output. Defaults to FALSE.}
}
\value{
A Seurat object containing the simulated gene expression data with
metadata and cell identifiers
}
\description{
This function simulates gene expression data using the Splatter package, which generates synthetic single-cell RNA-seq data based on a specified set of parameters.
The function allows for customization of the number of genes, cells, simulation method, group probabilities, and random seed for reproducibility.
The resulting simulated data is then converted into a Seurat object with appropriate metadata and cell identifiers.
}
\details{
Splatter doesn't use a separator for its fake names.
}
