% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{theme_bw_custom}
\alias{theme_bw_custom}
\title{ggplot2 theme with horizontal grid lines only}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.
}
\usage{
theme_bw_custom
}
\description{
ggplot2 theme with horizontal grid lines only
}
\keyword{datasets}
