% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vln_feat_plot}
\alias{vln_feat_plot}
\title{Plots a Seurat \code{VlnPlot} and a \code{FeaturePlot} side by side for the same marker}
\usage{
vln_feat_plot(
  seurat_obj,
  feature,
  assay = "RNA",
  group_col = NULL,
  rotate = FALSE
)
}
\arguments{
\item{seurat_obj}{The Seurat object with GEX data.}

\item{feature}{The feature of interest.}

\item{assay}{The assay to search for the feature in.}

\item{group_col}{What to group by (uses the Idents by default).}

\item{rotate}{Rotate the labels or not.}
}
\value{
Two patchworked Seurat plots.
}
\description{
Plots a Seurat \code{VlnPlot} and a \code{FeaturePlot} side by side for the same marker
}
\details{
Will put the highest expressing cells on top for the latter.
}
