% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{automated_annotation}
\alias{automated_annotation}
\title{This function runs automated annotation using specified methods.}
\usage{
automated_annotation(
  seurat_obj,
  annotation_method,
  reference = "pbmcref",
  azimuth_assay = "RNA",
  azimuth_levels = c("l1", "l2", "l3")
)
}
\arguments{
\item{seurat_obj}{The Seurat object. Must be the path to a H5AD object if using CellTypist.}

\item{annotation_method}{Which method to use: c("Azimuth", "CellTypist")}

\item{reference}{Reference or model to use for prediction. Defaults to "pbmcref" (for Azimuth).}

\item{azimuth_assay}{Assay to use for Azimuth}

\item{azimuth_levels}{Levels to process for Azimuth e.g. c("l1", "l2", "l3")}
}
\value{
A data.frame with the annotations for each cell
}
\description{
This function runs automated annotation using specified methods.
}
\details{
Supports Azimuth and CellTypist annotation methods.
Assumes that the \code{Cells()} of \code{seurat_obj} are properly formatted (i.e. unique).
For Azimuth with a Seurat v5 object, all of the layers have to be joined.
For CellTypist, assumes the H5AD file and predictions have already been generated.
}
