% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{automated_annotation}
\alias{automated_annotation}
\title{Run automated cell type annotation}
\usage{
automated_annotation(seurat_obj, annotation_method, reference = "pbmcref")
}
\arguments{
\item{seurat_obj}{The Seurat object. Must be the path to a H5AD object if using CellTypist.}

\item{annotation_method}{Which method to use: CellTypist"}

\item{reference}{Reference or model to use for prediction.}
}
\value{
A data.frame with the annotations for each cell
}
\description{
This function runs automated cell type annotation using \code{CellTypist}.
}
\details{
Supports CellTypist annotation methods.
Assumes that the \code{Cells()} of \code{seurat_obj} are properly formatted (i.e. unique).
For \code{CellTypist}, assumes the H5AD file and predictions have already been generated.
}
