% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airr.R
\name{bin_mu_freq}
\alias{bin_mu_freq}
\title{Bin the mutation frequency}
\usage{
bin_mu_freq(seurat_obj, num_bins = c(2, 3, 5))
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{num_bins}{The number of bins to split \code{mu_freq} into. Must be at least one of 2, 3, or 5.}
}
\value{
The provided Seurat object with a new binned mu_freq column.
}
\description{
Bins the \code{mu_freq} column in a Seurat object into specified categories (e.g. 0\%, 0-1\%, 1-5\%, etc.) for easier visualization and analysis.
The function creates new columns with binned mutation frequencies based on the provided number of bins.
}
\details{
The bins are (most likely) not going to be equal sizes.
}
