% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_adt_dists}
\alias{calc_adt_dists}
\title{Compute per-cell mean ADT distance to its k nearest neighbors from a base assay}
\usage{
calc_adt_dists(
  seurat_obj,
  base_assay,
  adt_assay = "ADT",
  layer = "data",
  feature,
  k,
  multiple_k = TRUE,
  distance_metric = "mean_abs",
  return_mean = TRUE,
  exclude_self = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object that contains neighbor slots for different k's.}

\item{base_assay}{Which neighbor space to use (e.g., "RNA", "BCR", "w").}

\item{adt_assay}{ADT assay name (ADT, ADTnorm).}

\item{layer}{Data layer to pull from (data, counts, or scale.data).}

\item{feature}{Optional vector of ADT features; if missing, use all features present.}

\item{k}{Number of nearest neighbors.}

\item{multiple_k}{Whether to look for a neighbor slot specific to the provided k (e.g., "RNA.nn_20") or just use the generic one (e.g., "RNA.nn"). The former allows you to have multiple neighbor graphs with different k's, while the latter assumes you only have one neighbor graph per assay.}

\item{distance_metric}{One of \code{c(mean_abs, manhattan, euclidean, cosine)}.}

\item{return_mean}{If TRUE, return the mean across all cells; else return per-cell.}

\item{exclude_self}{Drop the cell itself from neighbors if present.}
}
\value{
A single numeric value if \code{return_mean = TRUE}, or a named numeric vector of per-cell distances if \code{return_mean = FALSE}.
}
\description{
Compute per-cell mean ADT distance to its k nearest neighbors from a base assay
}
\details{
Pearson correlation coefficient measures linear relationships.
Euclidean distance measures the "straight-line" distance.
Cosine similarity measures the cosine of the angle between two vectors, with values closer to 1 meaning that they are more similar.
}
