% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_adt_scores}
\alias{calc_adt_scores}
\title{Calculate homogeneity scores for binary ADT features across embeddings}
\usage{
calc_adt_scores(
  seurat_objs,
  meta_res,
  metric_type,
  metrics,
  adt_features = "CD27.1",
  adt_cutoff = 1
)
}
\arguments{
\item{seurat_objs}{List of Seurat objects for each embedding type.}

\item{meta_res}{Named list of cluster columns for each reduction.}

\item{metric_type}{One of: Internal, External}

\item{metrics}{List of metric names.}

\item{adt_features}{ADT feature name (e.g. "CD27.1"). You can provide multiple.}

\item{adt_cutoff}{Numeric cutoff for binary classification.}
}
\value{
Data frame of scores for each embedding/reduction.
}
\description{
This function calculates homogeneity scores for binary ADT features across different embeddings and reductions in a Seurat object.
}
