% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_distances}
\alias{calc_distances}
\title{Calculate cluster distances for a Seurat object}
\usage{
calc_distances(
  seurat_obj,
  reduction_name,
  criteria = "Within_Max",
  labels_true = "annotated_clusters",
  labels_name
)
}
\arguments{
\item{seurat_obj}{The post-WNN Seurat object.}

\item{reduction_name}{The name of the reduction to use for distances.}

\item{criteria}{One of: Between_Mean, Between_Min, Calinski_Harabasz, Within_Between, Within_Max, Within_Mean, Within_Median.}

\item{labels_true}{The name of the column in the metadata that contains the cluster labels to evaluate.}

\item{labels_name}{A more descriptive name for the labels to use in plotting (optional).}
}
\value{
A data.frame with a row per metric and cluster containing the score.
}
\description{
This function calculates various cluster distance metrics using the \code{fpc::cluster.stats} function, which provides a comprehensive set of clustering statistics based on a distance matrix and cluster assignments.
The function computes both single-value metrics (e.g., Calinski-Harabasz index) and cluster-wise metrics (e.g., average within-cluster distance) depending on the specified criteria.
The results are returned in a tidy data frame format for easy plotting and comparison across different embeddings and reductions.
}
\details{
The \code{as.factor()} is needed in case you give a categorical cluster col.
Assumes you're using the embeddings approach.
}
