% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calc_moran}
\alias{calc_moran}
\title{Calculate Moran's i for a Seurat object}
\usage{
calc_moran(seurat_obj, feature, graph_name, row_standardize = TRUE)
}
\arguments{
\item{seurat_obj}{The Seurat object. Must have \code{FindNeighbors()} already run and an assay named "ADT".}

\item{feature}{Name of the ADT feature to evaluate (e.g. "CD27.1", "CD38").}

\item{graph_name}{Name of the neighbor graph slot to use for the weights matrix (e.g. "RNA.nn", "BCR.nn", "w.nn").}

\item{row_standardize}{Whether or not to row-standardize the weights matrix (i.e. make each row sum to 1).}
}
\value{
A single numeric value representing the observed Moran's i index for the specified feature and neighbor graph.
}
\description{
This function calculates the global Moran's i index.
}
\details{
We are using \code{MERINGUE}'s implementation instead of \code{ape}'s because it runs faster.
However, \code{MERINGUE} is not on CRAN, which means this package could not be published on CRAN.
Row standardization makes sure that the resulting score will always be between -1 and 1.
}
