% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassification.R
\name{calc_nn_frac}
\alias{calc_nn_frac}
\title{Calculate the proportion of a cell's neighbors that are mutated or unmutated}
\usage{
calc_nn_frac(seurat_obj, assay = "WNN", plot_title = "", category = "mutated")
}
\arguments{
\item{seurat_obj}{A Seurat object with neighbors calculated.}

\item{assay}{One of: RNA, BCR, WNN}

\item{plot_title}{A string to add to the plot title, e.g. "All Cells" or "Memory B Cells Only".}

\item{category}{One of: Mutated, Unmutated}
}
\value{
A data.frame and a plot.
}
\description{
This function calculates the proportion of neighbors that are mutated or unmutated for each cell in a Seurat object.
It uses the neighbor information stored in the Seurat object to determine the mutation status of neighboring cells and computes the fraction accordingly.
The function also generates a plot to visualize the distribution of mutated or unmutated neighbors across different cell types.
}
\details{
Assumes that WNN neighbors are stored in "w.nn".
Defines mutated as anything about 0\% SHM.
\code{setNames()} works better than a for loop checking all of the neighbors.
}
