% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{cell_type_clusters}
\alias{cell_type_clusters}
\title{Map cell types to Seurat clusters}
\usage{
cell_type_clusters(
  seurat_obj,
  clusters_col = "seurat_clusters",
  annotations_col
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{clusters_col}{The metadata column with the Seurat clusters.}

\item{annotations_col}{The metadata column with the cell types.}
}
\value{
A data.frame with a row for each cell type.
}
\description{
This function maps cell types to Seurat clusters by counting the number of cells in each cluster that belong to each cell type, and then assigning the most common cell type to each cluster.
It returns a data frame with the assigned cell types for each cluster.
}
\details{
This would typically be used after \code{\link[=seurat_pipeline]{seurat_pipeline()}} and \code{\link[=add_annotations]{add_annotations()}}.
}
