% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airr.R
\name{convert_embeddings}
\alias{convert_embeddings}
\title{Convert the output of an embedding method to a matrix}
\usage{
convert_embeddings(
  embeddings,
  combined_airr,
  combined_airr_input,
  verbose = TRUE
)
}
\arguments{
\item{embeddings}{The data.frame embeddings output.}

\item{combined_airr}{The combined output from airrflow/Immcantation. Must contain columns called "cell_id" and "cell_id_original".}

\item{combined_airr_input}{The data.frame provided to immune2vec; contains translated sequences.}

\item{verbose}{Logical indicating whether or not to print messages.}
}
\value{
A \code{Matrix} of embeddings
}
\description{
This function takes the output from an embedding method (e.g. AntiBERTy, ESM2, immune2vec) and converts it into a \code{Matrix} matrix format that can be used for downstream analysis.
It handles the mapping of cell IDs from the embeddings to the combined AIRR data, ensuring that only cells present in both datasets are retained.
The resulting matrix has features as rows and cells as columns, with appropriate column names based on the number of dimensions in the embeddings.
}
\details{
Assume that all inputs can be provided as tsv files.
The AMULETy outputs always have a column named "cell_id".
immune2vec does not include the cell ids in the output.
Uses \code{Matrix()} instead of \code{as.matrix()} since the latter only returns dense matrices.
AntiBERTy runs on 512 dimensions, AntiBERTa2 and BALM-paired run on 1024 dimensions, and ESM2 runs on 1280 dimensions (through AMULETy).
immune2vec is usually run with 100 dimensions.
}
