% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{gen_dot_title}
\alias{gen_dot_title}
\title{Generate a title for a \code{DotPlot}}
\usage{
gen_dot_title(plot_title = "", marker_sources)
}
\arguments{
\item{plot_title}{Dataset description}

\item{marker_sources}{The list of marker sources}
}
\value{
A string with sources comma-separated and in parentheses
}
\description{
This function generates a title for a DotPlot by combining a provided dataset description with a formatted list of marker sources.
The marker sources are sorted, formatted to replace underscores with spaces, and concatenated into a comma-separated string enclosed in parentheses.
}
