% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{get_airr_genes}
\alias{get_airr_genes}
\title{Get IG and TR genes from Ensembl using biomaRt.}
\usage{
get_airr_genes(
  genome = "hsapiens",
  ensembl_version = NULL,
  category = c("IG", "TR")
)
}
\arguments{
\item{genome}{The genome to use for gene annotation (e.g. "hsapiens" or "mmusculus").}

\item{ensembl_version}{The Ensembl version to use for gene annotation (e.g. "114").}

\item{category}{The category of genes to retrieve: "IG" for immunoglobulin genes, "TR" for T cell receptor genes, or both.}
}
\value{
A character vector of IG and/or TR gene names to be filtered out from the most variable features.
}
\description{
This function retrieves IG and TR genes from Ensembl using the biomaRt package. This allows for accurate IG and TR gene names instead of using a search for genes that begin with "IG" or "TR".
}
\details{
This will usually just be used as part of \code{\link[=seurat_pipeline]{seurat_pipeline()}}.
}
