% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{get_cell_types}
\alias{get_cell_types}
\title{Gets the information needed using \code{add_info_bar()} on a \code{DotPlot}}
\usage{
get_cell_types(markers_df)
}
\arguments{
\item{markers_df}{The markers data.frame filtered to match your input features}
}
\value{
A data.frame with Cell_Type_Full and features.plot columns
}
\description{
This function takes the filtered markers dataframe and selects the relevant columns to be used for adding an info bar to a DotPlot.
It renames the Marker column to features.plot and formats the Cell_Type_Full column for better readability.
}
