% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{labels_standard_vln}
\alias{labels_standard_vln}
\title{Standard label sizes for violin plots}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 6.
}
\usage{
labels_standard_vln
}
\description{
Standard label sizes for violin plots
}
\keyword{datasets}
