% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{labels_standard_vln_rotate}
\alias{labels_standard_vln_rotate}
\title{Violin plot labels with horizontal x-axis text}
\usage{
labels_standard_vln_rotate
}
\description{
Violin plot labels with horizontal x-axis text
}
