% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{named_colors}
\alias{named_colors}
\title{Named color palettes for athanor plots}
\format{
An object of class \code{list} of length 25.
}
\usage{
named_colors
}
\description{
A list of named character vectors mapping categorical values to hex colors,
used as the \code{clrs_specific} argument in plotting functions.
}
\keyword{datasets}
