% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{named_colors}
\alias{named_colors}
\title{Named color palettes for athanor plots}
\usage{
named_colors
}
\description{
A list of named character vectors mapping categorical values to hex colors,
used as the \code{clrs_specific} argument in plotting functions.
}
