% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{plot_counts_cluster}
\alias{plot_counts_cluster}
\title{Visualize how many cells are in each Seurat cluster or cell type}
\usage{
plot_counts_cluster(
  seurat_obj,
  tissue_type = "",
  clrs_specific,
  clusters_col,
  fill_col,
  fill_col_name,
  x_axis = "Cluster",
  details = NULL
)
}
\arguments{
\item{seurat_obj}{The Seurat object containing the clusters and/or cell type annotations to plot.}

\item{tissue_type}{The tissue type of interest e.g. "Blood" or "Skin".}

\item{clrs_specific}{The specific color palette (should be named).}

\item{clusters_col}{The column to plot on the x axis (e.g. seurat_clusters).}

\item{fill_col}{The column to fill by (e.g. annotated_clusters).}

\item{fill_col_name}{The label for the fill aesthetic.}

\item{x_axis}{What to plot on the x axis: "Cluster" or "Cell Type".}

\item{details}{The optional subtitle.}
}
\value{
A ggplot bar plot of cell counts.
}
\description{
This function generates a bar plot showing the number of cells in each Seurat cluster or cell type, optionally filled by another annotation (e.g. annotated clusters).
It allows for customization of the x-axis, fill colors, and plot titles.
}
