% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dimplot}
\alias{plot_dimplot}
\title{Plot a Seurat UMAP using \code{DimPlot}}
\usage{
plot_dimplot(
  seurat_obj,
  data_source = "",
  clrs_specific,
  use_hues = FALSE,
  pt_size = 0.2,
  title,
  reduc = "rna.umap",
  meta_col = "annotated_clusters",
  highlight,
  plot_label = FALSE,
  label_size = 3,
  label_box = TRUE,
  include_legend = TRUE,
  legend_label,
  sort_idents = TRUE,
  idents_char = TRUE,
  order = FALSE,
  details,
  ...
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{data_source}{Dataset description.}

\item{clrs_specific}{The specific color palette (should be named).}

\item{use_hues}{Use the iwanthue hues instead of the default ggplot colors. Doesn't let you set any other settings.}

\item{pt_size}{The point size.}

\item{title}{The plot title.}

\item{reduc}{The reduction to use for plotting e.g. "bpca" or wnn.umap".}

\item{meta_col}{Which column in the object metadata to color by. When combined with \code{highlight}, highlights those values as an overlay instead of coloring all cells.}

\item{highlight}{Can overlay clusters of interest e.g. B cell or by #. Overrides the annotated option.}

\item{plot_label}{Add labels to the plot (or not).}

\item{label_size}{The size of the plot labels.}

\item{label_box}{Whether or not to give the labels a background.}

\item{include_legend}{Include the legend or not.}

\item{legend_label}{The label for the legend.}

\item{sort_idents}{Whether or not to sort the idents (for proper ordering of the colors). This can mess up the order you want, so be careful.}

\item{idents_char}{If sorting idents, whether to sort them as characters or numerically (e.g. cluster 10 should be after cluster 9, not before).}

\item{order}{Plot cells on top or not.}

\item{details}{An optional custom subtitle.}

\item{...}{Any other Seurat parameters.}
}
\value{
A Seurat plot of the specified reduction.
}
\description{
This function generates a UMAP plot from a Seurat object using \code{DimPlot} with various customizable options for coloring, labeling, and grouping the data.
}
