% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mws}
\alias{plot_mws}
\title{Plot a box plot of modality weights per cell type}
\usage{
plot_mws(
  seurat_obj,
  details = "",
  second_assay = "BCR",
  clrs_specific = named_colors$mu_freq_bins,
  split_by = "mu_freq_bins",
  y_axis_label = "SHM Frequency Bins"
)
}
\arguments{
\item{seurat_obj}{The post-WNN Seurat object.}

\item{details}{Details to add to the plot title.}

\item{second_assay}{List of other assays run through WNN in order.}

\item{clrs_specific}{A specific (must have names) color palette.}

\item{split_by}{A meta.data column to split the box plots up by.}

\item{y_axis_label}{Label for the y-axis.}
}
\value{
A ggplot with the distribution of weights
}
\description{
This function creates a box plot to visualize the distribution of modality weights (e.g. RNA vs. BCR) across different cell types or clusters in a post-WNN Seurat object.
It should be run after \code{\link[=run_wnn]{run_wnn()}}.
}
\details{
Assumes "annotated_clusters" is a column.
}
