% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_dt}
\alias{print_dt}
\title{Displays a sortable, scrollable DataTable table when the R Markdown file knits.}
\usage{
print_dt(table, dt_width = "800px")
}
\arguments{
\item{table}{The input data frame to be printed.}

\item{dt_width}{The width of the output table.}
}
\value{
A DataTable table.
}
\description{
Displays a sortable, scrollable DataTable table when the R Markdown file knits.
}
