% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce_object}
\alias{reduce_object}
\title{Reduce a Seurat object's size}
\usage{
reduce_object(
  seurat_obj,
  dim_reducs,
  meta_cols,
  remove_neighbors = TRUE,
  print_size = TRUE,
  ...
)
}
\arguments{
\item{seurat_obj}{Processed Seurat object to reduce.}

\item{dim_reducs}{Vector of dimensionality reductions to keep.}

\item{meta_cols}{Vector of metadata column names to keep. If unspecified, keeps all metadata columns.}

\item{remove_neighbors}{Whether or not to remove neighbor graphs from the object.}

\item{print_size}{Whether to print info about how much the object was reduced.}
}
\value{
A reduced Seurat object with specified reductions kept.
}
\description{
This function reduces a Seurat object by removing count matrices and keeping only
specified dimensionality reductions. This is especially useful for creating
lightweight objects for Shiny apps or sharing.
}
\details{
Modify this as needed if your object is built differently (e.g. tSNE instead).
Uses \code{DietSeurat} to remove counts while preserving reductions and metadata.
This is especially useful if you are making a Shiny app or just visualizing the data.
This works well in conjunction with \code{\link[=seurat_pipeline]{seurat_pipeline()}}.
}
