% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{regen_reduc}
\alias{regen_reduc}
\title{Regenerate neighbor graphs and UMAPs.}
\usage{
regen_reduc(
  seurat_obj,
  pca_name = "rpca",
  assay = "RNA",
  num_dims = 20,
  k_param = 20,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{The Seurat object.}

\item{pca_name}{Name of the PCA reduction to use for neighbor finding and UMAP. This should be the name of an existing PCA reduction in the Seurat object (e.g. "pca" or "rpca").}

\item{assay}{Name of the assay to use for neighbor finding and UMAP.}

\item{num_dims}{Number of PCA dimensions to use for neighbor finding.}

\item{k_param}{Number of nearest neighbors.}

\item{verbose}{Print out Seurat's progress messages.}
}
\value{
A processed Seurat object with the \code{graphs}, \code{neighbors}, and \code{reductions} slots filled in or updated.
}
\description{
Regenerates PCA, neighbor graphs, and UMAP for a Seurat object.
This is useful if you subset a Seurat object and need to recompute these reductions for the new object.
It can also be used to regenerate these reductions if you have modified the object in a way that requires them to be redone (e.g. filtering out cells or genes).
}
