% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{seurat_pipeline}
\alias{seurat_pipeline}
\title{Runs Seurat's clustering pipeline with given parameters.}
\usage{
seurat_pipeline(
  seurat_obj,
  nfeatures_RNA = 200,
  perc_mt = 15,
  num_features = 2000,
  num_pcs = 30,
  num_dims = 20,
  cluster_res = 0.4,
  filter_genes = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{The Seurat object containing combined GEX data.}

\item{nfeatures_RNA}{Minimum number of RNA features.}

\item{perc_mt}{Maximum percentage of mitochondrial genes to retain.}

\item{num_features}{Desired number of variable features.}

\item{num_pcs}{Number of principal components to compute.}

\item{num_dims}{Number of dimensions to use for neighbor finding and UMAP.}

\item{cluster_res}{Clustering resolution parameter.}

\item{filter_genes}{Whether to filter out IG/TR genes.}

\item{verbose}{Print out Seurat's progress messages.}
}
\value{
A processed Seurat object with normalization, scaling, PCA, clustering, and UMAP.
}
\description{
Runs Seurat's clustering pipeline with given parameters.
}
\details{
It is highly recommended to save the resulting object as an RDS or qs file.
For filter_genes, we assume that \code{features_meta} is already loaded and \code{remove_genes} defined.
This pipeline is loosely based on \href{https://satijalab.org/seurat/articles/pbmc3k_tutorial}{Seurat's pipeline}.
Unlike previous analyses, \code{features = rownames(obj)} was removed from the \code{ScaleData} step since the data is too large and only the top variable features are needed to do \code{RunPCA}.
}
