% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex.R
\name{seurat_pipeline}
\alias{seurat_pipeline}
\title{Run Seurat's standard pipeline}
\usage{
seurat_pipeline(
  seurat_obj,
  nfeatures_RNA = 200,
  perc_mt = 15,
  num_features = 2000,
  num_pcs = 30,
  num_dims = 20,
  cluster_res = 0.4,
  filter_genes,
  ensembl_version = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{The Seurat object containing combined GEX data.}

\item{nfeatures_RNA}{Minimum number of RNA features.}

\item{perc_mt}{Maximum percentage of mitochondrial genes to retain.}

\item{num_features}{Desired number of variable features.}

\item{num_pcs}{Number of principal components to compute.}

\item{num_dims}{Number of dimensions to use for neighbor finding and UMAP.}

\item{cluster_res}{Clustering resolution parameter.}

\item{filter_genes}{If specified, filter out genes from this category (e.g. "IG" and/or "TR")}

\item{verbose}{Print out Seurat's progress messages.}
}
\value{
A processed Seurat object with normalization, scaling, PCA, clustering, and UMAP.
}
\description{
This function takes a Seurat object containing combined GEX data and runs the standard Seurat pipeline for normalization, scaling, dimensionality reduction, clustering, and UMAP visualization.
It includes optional filtering of IG/TR genes and allows for customization of various parameters such as the number of variable features, principal components, and clustering resolution.
}
\details{
It is highly recommended to save the resulting object as an RDS or qs file.
This pipeline is loosely based on \href{https://satijalab.org/seurat/articles/pbmc3k_tutorial}{Seurat's pipeline}.
Unlike previous analyses, \code{features = rownames(obj)} was removed from the \code{ScaleData} step since the data is too large and only the top variable features are needed to do \code{RunPCA}.
}
