% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{sim_bcr_manual}
\alias{sim_bcr_manual}
\title{This simulates an assay made from BCR embeddings independently of any GEX objects}
\usage{
sim_bcr_manual(num_cells, num_dims, separator)
}
\arguments{
\item{num_cells}{Number of cells to simulate.}

\item{num_dims}{Number of embedding dimensions to simulate.}

\item{separator}{Separator for cell and dimension names.}
}
\value{
A Seurat Assay
}
\description{
This simulates an assay made from BCR embeddings independently of any GEX objects
}
\details{
The choice of value ranges was based off the real COVID-19 data, including having 9 decimal points
}
